Add system- and user-space feature flagging ability
Flags are defined in public/feature_flags.json with a name,
system-level enable flag and user-level overriding settings.
They are stored in Redux and polled regularly in case of server-side
updates. Components or logic can query the Redux store as usual to
make decisions based on feature flag state: useFeatureFlag hook or
selectIsFeatureFlagEnabled selector.
At compile time, the strings used as flags are checked against the
JSON file (by importing it as a constant and making a type from it)
so typos should fail a build. At run time unknown flags default to
disabled.

A flag without (system-level) enabled="true" disables its features
entirely for all users (on that deployment).
User overrides allow users to pick and choose their feature set from
those flags which are marked as user-configurable. The user may
choose between turning a feature on, off or following the system-wide
default_enabled property.
Any flag that is disabled at the system level cannot be overridden
at the user level.

Flags can declare dependencies on other flags via depends_on. Circular
dependencies are detected and result in the flag being disabled.

Care should be taken when deploying to not override manual changes
in the feature_flags.json and if deploying a rollback to add a manual
disable for any newly-added features (for any user who has not refreshed
to the rollback version).